backend/ai/ai_socket/requirements.txt
openai
python-dotenv
requests
django
djangorestframework
django-cors-headers